# Ensuring the integrity of full snapshot before uploading it to the object store.