**What this PR does / why we need it**:
It has been observed that while doing the restoration from full snapshot sometimes backup-restore failed to restore the etcd due to sometimes full snapshots got corrupted or missing a hash.
This PR is try to minimise the occurrence of such scenarios by verifying the integrity of full snapshot before uploading it to the object store.

**Which issue(s) this PR fixes**:
Fixes #778 

**Special notes for your reviewer**:
This is how it's done:

<img width="893" alt="Screenshot 2024-09-23 at 10 10 45 AM" src="https://github.com/user-attachments/assets/c30cdf4d-b0b1-423a-b932-77eb878cc780">


**Release note**:
<!--  Write your release note:
1. Enter your release note in the below block.
2. If no release note is required, just write "NONE" within the block.

Format of block header: <category> <target_group>
Possible values:
- category:       improvement|noteworthy|action
- target_group:   user|operator|developer
-->
```noteworthy operator
Ensure the integrity of the full snapshot prior to uploading it to the object store, thereby reducing the potential restoration failures.
```