@seshachalam-yv 2024-10-21
**Benchmark Test Results Comparison Between PR Branch and Master Branch**

I conducted benchmark tests to compare the performance of the `TakeAndSaveFullSnapshot` function between the current PR branch and the master branch.

**Master Branch (`old.txt`):**
```
goos: darwin
goarch: arm64
pkg: github.com/gardener/etcd-backup-restore/pkg/etcdutil
cpu: Apple M1 Max
BenchmarkTakeAndSaveFullSnapshot-10          100     466519650 ns/op    252784125 B/op    29873 allocs/op
PASS
ok      github.com/gardener/etcd-backup-restore/pkg/etcdutil   47.593s
```

**PR Branch (`new.txt`):**
```
goos: darwin
goarch: arm64
pkg: github.com/gardener/etcd-backup-restore/pkg/etcdutil
cpu: Apple M1 Max
BenchmarkTakeAndSaveFullSnapshot-10          100     571449465 ns/op    256778608 B/op    30696 allocs/op
PASS
ok      github.com/gardener/etcd-backup-restore/pkg/etcdutil   58.196s
```

```
                         │     Old      │     New      │    Delta
-------------------------│--------------│--------------│-----------------
Execution Time (sec/op)  │ 466.5ms ± 0% │ 571.4ms ± 0% │ +22.5% (↑)
Memory Usage (B/op)      │ 241.1MiB ± 0%│ 244.9MiB ± 0%│  +1.57% (↑)
Allocations (allocs/op)  │ 29,873 ± 0%  │ 30,696 ± 0%  │  +2.78% (↑)
```
**Conclusion**

There is a 22.5% increase in execution time in the new version. However, since this increase amounts to approximately `105 milliseconds` per operation, it may not have a significant impact on overall performance.

WDYT @ishan16696